requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.10.0
asposestorage==1.0.2
audioread==3.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.3
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==10.4.0
platformdirs==4.4.0
pooch==1.8.2
psutil==7.0.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyworld==0.3.4
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomlkit==0.12.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.67.1
transformers==4.47.1
triton==2.3.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
utils==1.0.2
uvicorn==0.35.0
websockets==12.0
zipp==3.23.0
